anime streaming routes and controllers updated
